goircam: initialize the bluetooth after GO.begin()
On my Odroid-GO, the current firmware immediately hangs at boot.
It happens that calling serailBT.begin() before GO.begin() does
not work. There is possibly a race there in the init code though
because while doing this fixes the problem, sometimes it's also
required to add a delay(500) between the two and sometimes it
needs to be removed, depending on code placed elsewhere and
totally unrelated to this.