goircam: use floats for the temperatures
Temperatures were converted a bit too early to integers, resulting in
a big loss of precision in case the upper and lower ones were close to
each other. Let's perform the computation on floats instead, and use
1000*float as integer when needed (e.g. maps) since the thermal sensor
provides 18 bits of precision over the range, hence ~1mK resolution.
The lower, mid and upper temperatures are also reported as floats so
that it's easier to know the min and max.