goircam: place a delay before reading the joypad to avoid erratic readings
The joypad reading regularly reports erratic values making the box
occasionally move in the wrong direction (X2 often reported as X1,
and Y2 often reported as Y1). It appears that placing a delay(50)
before reading solves this, so this might be the result of some
noise before reading the inputs. Also Y1 was missing the debounce
loop.