goircam: keep the same temperature scale when not releasing the button
It's possible to take continuous snapshots by holding the A button, but
with a permanently adjusted temperature scale, colors constantly change
and this doesn't help to visually spot cold/hot places. This change makes
sure we only recompute the low/high range of the temperature scale when
the A button is not pressed. This way keeping it pressed will stick to
the same scale for all subsequent snapshots.