goircam: make the surrounding pixels count for up to 25% of the color
In order to display a smoother image and better take into account the
chessmode of the sensor, let's make the 4 closest surrounding pixels
count for up to 25% of the displayed pixel's color (i.e. each of them
counts for 6.25%). Since these 4 pixels come from the alternate plane
which is not scanned at the same time, it significantly reduces the
noise on the picture.